new_requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.9.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.4.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
toml==0.10.2
tzdata==2025.2
urllib3==1.26.20
